+++ activemodel/lib/active_model/validations/helper_methods.rb
@@ -1,0 +1,13 @@
+module ActiveModel
+  module Validations
+    module HelperMethods
+      private
+      def _merge_attributes(attr_names)
+        options = attr_names.extract_options!.symbolize_keys
+        attr_names.flatten!
+        options[:attributes] = attr_names
+        options
+      end
+    end
+  end
+end
